Pete dev - Refactor forum data types and session management - Breaking Schema Change

+++ source/digioz.Forum/digioz.Forum/Models/Forum.cs
@@ -55,7 +55,7 @@ public partial class Forum
 
     public string ForumLastPostSubject { get; set; } = null!;
 
-    public int ForumLastPostTime { get; set; }
+    public long ForumLastPostTime { get; set; }
 
     public string ForumLastPosterName { get; set; } = null!;
